updated glutin version + use std::f32::consts::PI

+++ src/state/mouse.rs
@@ -5,7 +5,7 @@ pub enum MouseButton {
     Left,
     Right,
     Middle,
-    Other(u8),
+    Other(u16),
 }
 
 #[derive(Copy, Clone, PartialEq, Debug)]
